+++ contracts/index_executor/src/steps/transfer.rs
@@ -87,7 +87,7 @@ impl Runner for TransferStep {
 
         // Query off-chain indexer directly get the execution result
         let chain = &context
-            .graph
+            .registry
             .get_chain(self.chain.clone())
             .ok_or("MissingChain")?;
         let account = match chain.chain_type {
